Add files via upload
AGRAND is an innovative agriculture technology solution designed to simplify and enhance the farming experience. This demo showcases the working model of the application, providing farmers with a glimpse into the powerful features and capabilities of the platform. From crop yield prediction to weather tracking and soil analysis, AGRAND has it all. Experience the future of farming today with AGRAND."